update3

+++ EntityLayer/Uye.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -14,7 +15,7 @@ namespace EntityLayer
         public string Soyad { get; set; }
         public string Eposta { get; set; }
         public string Sifre { get; set; }
-        public int Telefon { get; set; }
+        public double Telefon { get; set; }
         public int? SehirId { get; set; }
         public int? IlceId { get; set; }
         public virtual List<Bilet> Biletler { get; set; }
